Remove helpers and integrate logic into controllers
