+++ root/app/Controllers/InfoController.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace App\Controllers;
 
 use App\Core\Controller;
@@ -9,7 +10,7 @@ class InfoController extends Controller
 {
     public static function handleRequest(): void
     {
-        AuthMiddleware::checkSession();
+        AuthMiddleware::check();
 
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             if (!isset($_POST['csrf_token']) || $_POST['csrf_token'] !== $_SESSION['csrf_token']) {
@@ -64,5 +65,29 @@ class InfoController extends Controller
         }
 
         $this->render('info');
+    }
+
+    public static function generateProfileDataAttributes(string $username): string
+    {
+        $userInfo = UserHandler::getUserInfo($username);
+        if ($userInfo) {
+            $data = "data-who=\"" . htmlspecialchars($userInfo->who) . "\" ";
+            $data .= "data-where=\"" . htmlspecialchars($userInfo->where) . "\" ";
+            $data .= "data-what=\"" . htmlspecialchars($userInfo->what) . "\" ";
+            $data .= "data-goal=\"" . htmlspecialchars($userInfo->goal) . "\"";
+            return $data;
+        }
+        return '';
+    }
+
+    public static function buildSystemMessage(string $username): string
+    {
+        $userInfo = UserHandler::getUserInfo($username);
+        if ($userInfo) {
+            $systemMessage = "<span style=\"color: blue; font-weight: bold;\">" . SYSTEM_MSG . "</span>";
+            $systemMessage .= " <span style=\"color: blue; font-weight: bold;\">You work for</span> " . htmlspecialchars($userInfo->who) . " <span style=\"color: blue; font-weight: bold;\">located in</span> " . htmlspecialchars($userInfo->where) . ". " . htmlspecialchars($userInfo->what) . " <span style=\"color: blue; font-weight: bold;\">Your goal is</span> " . htmlspecialchars($userInfo->goal) . ".";
+            return $systemMessage;
+        }
+        return "<span style=\"color: blue; font-weight: bold;\">" . SYSTEM_MSG . "</span>";
     }
 }
