+++ root/app/Views/accounts.php
@@ -1,5 +1,5 @@
 <?php
-use App\Helpers\AccountsHelper;
+use App\Controllers\AccountsController;
 
 require 'layouts/header.php';
 ?>
@@ -33,12 +33,12 @@ require 'layouts/header.php';
                 <!-- Days selection dropdown (multiple) -->
                 <label for="days">Days:</label>
                 <select class="form-select" name="days[]" id="days" multiple required>
-                    <?php echo AccountsHelper::generateDaysOptions(); ?>
+                    <?php echo AccountsController::generateDaysOptions(); ?>
                 </select>
                 <!-- Post schedule selection dropdown (multiple) -->
                 <label for="cron">Post Schedule:</label>
                 <select class="form-select" name="cron[]" id="cron" multiple required>
-                    <?php echo AccountsHelper::generateCronOptions(); ?>
+                    <?php echo AccountsController::generateCronOptions(); ?>
                 </select>
                 <!-- Hashtags inclusion dropdown -->
                 <label for="hashtags">Include Hashtags:</label>
@@ -56,7 +56,7 @@ require 'layouts/header.php';
                 <h3 class="account-name card-title">Account List</h3>
             </div>
             <div class="columns">
-                <?php echo AccountsHelper::generateAccountList(); ?>
+                <?php echo AccountsController::generateAccountList(); ?>
             </div>
         </div>
     </div>
